+++ android/src/main/kotlin/com/adyen/checkout/flutter/components/view/DynamicComponentView.kt
@@ -10,7 +10,7 @@ import androidx.appcompat.widget.SwitchCompat
 import androidx.core.content.ContextCompat
 import androidx.core.view.children
 import androidx.core.view.postDelayed
-import com.adyen.checkout.card.CardComponent
+import com.adyen.checkout.components.core.internal.ButtonComponent
 import com.adyen.checkout.components.core.internal.Component
 import com.adyen.checkout.flutter.generated.ComponentCommunicationModel
 import com.adyen.checkout.flutter.generated.ComponentCommunicationType
@@ -94,7 +94,7 @@ class DynamicComponentView
             adyenComponentView.getViewTreeObserver()?.addOnGlobalLayoutListener(
                 object : ViewTreeObserver.OnGlobalLayoutListener {
                     override fun onGlobalLayout() {
-                        if (component is CardComponent) {
+                        if (component is ButtonComponent) {
                             overrideSubmit(component)
                         }
 
@@ -104,7 +104,7 @@ class DynamicComponentView
             )
         }
 
-        private fun overrideSubmit(component: CardComponent) {
+        private fun overrideSubmit(component: ButtonComponent) {
             val payButton = findViewById<MaterialButton>(com.adyen.checkout.ui.core.R.id.payButton)
             if (android.os.Build.VERSION.SDK_INT <= android.os.Build.VERSION_CODES.O) {
                 disableRippleAnimationOnPayButton()