+++ ios/adyen_checkout/Sources/adyen_checkout/utils/AdyenAppearance.swift
@@ -6,6 +6,7 @@ import Adyen
 public enum AdyenAppearance {
     private static var _dropInStyle: DropInComponent.Style = .init()
     private static var _cardComponentStyle: Adyen.FormComponentStyle = .init()
+    private static var _blikComponentStyle: Adyen.FormComponentStyle = .init()
 
     public static var dropInStyle: DropInComponent.Style {
         get { _dropInStyle }
@@ -15,5 +16,10 @@ public enum AdyenAppearance {
     public static var cardComponentStyle: Adyen.FormComponentStyle {
         get { _cardComponentStyle }
         set { _cardComponentStyle = newValue }
+    }
+    
+    public static var blikComponentStyle: Adyen.FormComponentStyle {
+        get { _blikComponentStyle }
+        set { _blikComponentStyle = newValue }
     }
 }